Use SOCK_RAW and cast iov into ip and icmphdr
With SOCK_RAW we get the ip layer too, which means we have access
to all the info we need from the echo reply packet (ttl for example).

The source address is in the header so we should not need msg_name.
Test if it works if we delete it.